# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EdgeEver API",
    "version": "0.1.0",
    "description": "REST API for EdgeEver notebooks, memos, resources, API tokens, and Agent integrations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/ai/settings": {
      "get": {
        "summary": "Read the current workspace AI model settings",
        "responses": {
          "200": {
            "description": "AI model settings with the API key omitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "settings": { "anyOf": [{ "$ref": "#/components/schemas/AiModelSettings" }, { "type": "null" }] },
                    "encryptionConfigured": { "type": "boolean" }
                  },
                  "required": ["settings", "encryptionConfigured"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      },
      "put": {
        "summary": "Save the current workspace AI model settings",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiModelSettingsInput" } } }
        },
        "responses": {
          "200": { "description": "AI model settings saved" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/settings/test": {
      "post": {
        "summary": "Test an AI model endpoint without exposing its credential",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiModelSettingsInput" } } }
        },
        "responses": {
          "200": { "description": "The model endpoint responded successfully" },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/api/v1/ai/generate": {
      "post": {
        "summary": "Stream a note AI action using the current workspace model",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": { "type": "string", "enum": ["summarize", "extract-key-points", "extract-todos", "rewrite-proofread", "translate"] },
                  "title": { "type": "string", "maxLength": 160 },
                  "contentMarkdown": { "type": "string", "maxLength": 300000 },
                  "targetLanguage": { "type": "string", "maxLength": 80, "description": "Required when action is translate" }
                },
                "required": ["action", "title", "contentMarkdown"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-sent events containing start, text-delta, finish, or error events",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" },
          "409": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/v1/api-tokens": {
      "get": {
        "summary": "List API tokens",
        "responses": {
          "200": {
            "description": "API token list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiTokens": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApiToken"
                      }
                    },
                    "availableScopes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": ["apiTokens", "availableScopes"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create API token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiTokenCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created token; plaintext token is shown once",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiToken"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/api-tokens/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ApiTokenId"
        }
      ],
      "delete": {
        "summary": "Revoke API token",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "edgeever_session"
      }
    },
    "parameters": {
...
      "ApiTokenId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "Ok": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Ok"
            }
          }
        }
      },
      "Error": {
        "description": "API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": ["ok"]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "AiModelSettings": {
        "type": "object",
        "properties": {
          "provider": { "type": "string", "enum": ["openai-compatible", "anthropic", "google"] },
          "displayName": { "type": "string" },
          "baseUrl": { "type": "string", "format": "uri" },
          "modelId": { "type": "string" },
          "isEnabled": { "type": "boolean" },
          "hasApiKey": { "type": "boolean" },
          "encryptionConfigured": { "type": "boolean" }
        },
        "required": ["provider", "displayName", "baseUrl", "modelId", "isEnabled", "hasApiKey", "encryptionConfigured"]
      },
      "AiModelSettingsInput": {
        "type": "object",
        "properties": {
          "provider": { "type": "string", "enum": ["openai-compatible", "anthropic", "google"] },
          "displayName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "baseUrl": { "type": "string", "format": "uri", "maxLength": 500 },
          "apiKey": { "type": "string", "minLength": 1, "maxLength": 4096 },
          "modelId": { "type": "string", "minLength": 1, "maxLength": 200 },
          "isEnabled": { "type": "boolean", "default": true }
        },
        "required": ["provider", "displayName", "baseUrl", "modelId"]
      },
...
      "ApiToken": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "token": {
            "type": ["string", "null"]
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastUsedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "expiresAt": {
            "type": ["string", "null"],
            "format": "date-time"
          },
          "isRevoked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "name", "token", "scopes", "lastUsedAt", "expiresAt", "isRevoked", "createdAt"]
      },
      "ApiTokenCreateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "scopes": {
            "type": "array",
            "minItems": 1,
            "maxItems": 32,
            "items": {
              "type": "string"
            }
          },
          "expiresAt": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        },
        "required": ["name", "scopes"]
      },
      "CreatedApiToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "apiToken": {
            "$ref": "#/components/schemas/ApiToken"
          }
        },
        "required": ["token", "apiToken"]
      },
...
    }
  }
}